+++ src/routes/themes/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
+	const PlaceholderTitle = "Hello world, I'm theme";
 </script>
 
 <div class="flex justify-between items-center gap-4">
-	<h3>Hello World</h3>
+	<h3>{PlaceholderTitle}</h3>
 </div>
